// src/intor/core/intor-config/types/define-intor-config-types.ts
import type {
  InitCookieOptions,
  ResolvedCookieOptions,
} from "./cookie-options-types";
import type { IntorAdapter } from "./intor-adapter-types";
import type { LoaderOptions } from "./loader-options-types";
import type { InitLoggerOptions } from "./logger-options-types";
import type {
  InitRoutingOptions,
  ResolvedRoutingOptions,
} from "./routing-options-types";
import type {
  InitTranslatorOptions,
  ResolvedTranslatorOptions,
} from "./translator-options-types";
import type {
  FallbackLocalesMap,
  Locale,
  LocaleNamespaceMessages,
} from "../../../types/message-structure-types";
import type { PrefixPlaceHolder } from "../../../types/pathname-types";

// With static messages (Without loader options)
export type WithStaticMessages = {
  loaderOptions?: undefined;
  supportedLocales?: readonly Locale[]; // Optional
};

// With dynamic messages (With loader options)
export type WithDynamicMessages = {
  loaderOptions: LoaderOptions;
  supportedLocales: readonly Locale[]; // Required
};

// =====================================
// Init config
// =====================================

export type IntorInitConfig = (
  | WithDynamicMessages // supportedLocales is required
  | WithStaticMessages
) & {
  readonly id?: string; // Identifier for the intor instance
  readonly messages?: LocaleNamespaceMessages; // Static messages
  readonly defaultLocale: Locale;
  readonly fallbackLocales?: FallbackLocalesMap;
  readonly translator?: InitTranslatorOptions;
  readonly cookie?: InitCookieOptions;
  readonly routing?: InitRoutingOptions;
  readonly adapter?: IntorAdapter;
  readonly logger?: InitLoggerOptions;
  readonly prefixPlaceHolder?: PrefixPlaceHolder;
};

// =====================================
// Resolved config
// =====================================

export type IntorResolvedConfig = (
  | WithDynamicMessages // supportedLocales is required
  | WithStaticMessages
) & {
  readonly id: string; // Identifier for the intor instance
  readonly messages?: LocaleNamespaceMessages; // Static messages
  readonly defaultLocale: Locale;
  readonly fallbackLocales: FallbackLocalesMap;
  readonly translator: ResolvedTranslatorOptions;
  readonly cookie: ResolvedCookieOptions;
  readonly routing: ResolvedRoutingOptions;
  readonly adapter: IntorAdapter;
  readonly prefixPlaceHolder: PrefixPlaceHolder;
};